chore(core/modules): move game manager to game sub directory
