+++ apps/core/src/index.ts
@@ -11,7 +11,7 @@ import { checkVersion } from './common/version_check';
 import apiRoutes from './routes/api';
 import internalRoutes from './routes/internal';
 import { ProcessManager } from './modules/process.manager';
-import { GameManager } from './modules/game.manager';
+import { GameManager } from './modules/game/manager';
 import { ConfigManager } from './modules/config/manager';
 import { perfManager } from './modules/perf/manager';
 import { applyMigrations } from '@fxmanager/database';
