+++ apps/core/src/modules/game/manager.ts
@@ -8,9 +8,9 @@ import type {
 	PlayerIdentifiers,
 	PlayerUpdatePackage,
 } from '@fxmanager/shared/types';
-import { wsManager } from './ws.manager';
-import { discordManager } from './discord/manager';
-import { ConfigManager } from './config/manager';
+import { wsManager } from '../ws.manager';
+import { discordManager } from '../discord/manager';
+import { ConfigManager } from '../config/manager';
 
 export class GameManager {
 	private playerlist: OnlinePlayer[] = [];
